requirements.txt
# Discord & Bot logic
discord.py
python-dotenv

# Google Gemini API
google-genai

# RPG Scribe / PDF Processing
pypdf
pymupdf
pymupdf4llm[ocr,layout]

# Testing
pytest
pytest-asyncio

# Visualization
tqdm
